Reorganization part 999 for my model Keep note of the directory tree change! Setting up for use with OpenAI gym and keras-rl. To run minesweeper itself, call python FROM the ROOT directory of the project (the images are at the root).
Added features:
- "Cheat" keys:
	- F5 reveals all bombs
	- F6 reveals all non-bombs (dry reveal)
	- F7 resets the game
Unfinished as of this commit:
- The entire model so far...
- grid_space.py: will be a gym.Space object to represent Grid

Keep note that this may break anything you were working on if you have not committed it before this merge! The directory structure is not complex, if you have your solving algorithm anywhere, move it to apcspminesweeper/envs/ or `from apcspminesweeper.envs import Minesweeper` to use the Minesweeper object as normal.